refactor: response 관련 inner enum 바깥으로 분리
