+++ src/main/java/webserver/response/HttpResponse.java
@@ -62,50 +62,4 @@ public class HttpResponse {
             logger.error(e.getMessage());
         }
     }
-
-    public enum HttpResponseHeader {
-        CONTENT_TYPE("Content-Type"),
-        CONTENT_LENGTH("Content-Length"),
-        LOCATION("Location"),
-        ;
-
-        private final String header;
-
-        HttpResponseHeader(String header) {
-            this.header = header;
-        }
-
-        public String getHeader() {
-            return header;
-        }
-    }
-
-    public enum HttpResponseStatus {
-        OK("200 OK"),
-        REDIRECT("302 Found");
-
-        private final String status;
-
-        HttpResponseStatus(String status) {
-            this.status = status;
-        }
-
-        public String getStatus() {
-            return status;
-        }
-    }
-
-    public enum HttpVersion {
-        HTTP1_1("HTTP/1.1");
-
-        private final String httpVersion;
-
-        HttpVersion(String httpVersion) {
-            this.httpVersion = httpVersion;
-        }
-
-        public String getHttpVersion() {
-            return httpVersion;
-        }
-    }
 }
